+++ hanfor/tests/test_simulator/test_counter_trace.py
@@ -4,7 +4,7 @@ from lark.lark import Lark
 from parameterized import parameterized
 from pysmt.fnode import FNode
 from pysmt.shortcuts import Symbol, FALSE
-from pysmt.typing import INT
+from pysmt.typing import REAL
 
 from simulator.counter_trace import CounterTraceTransformer
 
@@ -22,12 +22,12 @@ testcases = {
          'True;True'),
 
     'true_lower_bound_empty':
-        ({'T': Symbol('T', INT)},
+        ({'T': Symbol('T', REAL)},
          'true ∧ ℓ >₀ T;true',
          'True ∧ ℓ >₀ T;True'),
 
     'true_lower_bound':
-        ({'T': Symbol('T', INT)},
+        ({'T': Symbol('T', REAL)},
          'true ∧ ℓ > T;true',
          'True ∧ ℓ > T;True'),
 
@@ -57,52 +57,52 @@ testcases = {
          'True;⌈P⌉;⌈(! Q)⌉;⌈((! Q) & R)⌉;True'),
 
     'duration_bound_l_globally':
-        ({'R': Symbol('R'), 'T': Symbol('T', INT)},
+        ({'R': Symbol('R'), 'T': Symbol('T', REAL)},
          'true;⌈!R⌉;⌈R⌉ ∧ ℓ < T;⌈!R⌉;true',
          'True;⌈(! R)⌉;⌈R⌉ ∧ ℓ < T;⌈(! R)⌉;True'),
 
     'duration_bound_l_before':
-        ({'P': Symbol('P'), 'R': Symbol('R'), 'T': Symbol('T', INT)},
+        ({'P': Symbol('P'), 'R': Symbol('R'), 'T': Symbol('T', REAL)},
          '⌈!P⌉;⌈(!P && !R)⌉;⌈(!P && R)⌉ ∧ ℓ < T;⌈(!P && !R)⌉;true',
          '⌈(! P)⌉;⌈((! P) & (! R))⌉;⌈((! P) & R)⌉ ∧ ℓ < T;⌈((! P) & (! R))⌉;True'),
 
     'duration_bound_l_after':
-        ({'P': Symbol('P'), 'R': Symbol('R'), 'T': Symbol('T', INT)},
+        ({'P': Symbol('P'), 'R': Symbol('R'), 'T': Symbol('T', REAL)},
          'true;⌈P⌉;true;⌈!R⌉;⌈R⌉ ∧ ℓ < T;⌈!R⌉;true',
          'True;⌈P⌉;True;⌈(! R)⌉;⌈R⌉ ∧ ℓ < T;⌈(! R)⌉;True'),
 
     'duration_bound_l_between':
-        ({'P': Symbol('P'), 'Q': Symbol('Q'), 'R': Symbol('R'), 'T': Symbol('T', INT)},
+        ({'P': Symbol('P'), 'Q': Symbol('Q'), 'R': Symbol('R'), 'T': Symbol('T', REAL)},
          'true;⌈(P && !Q)⌉;⌈!Q⌉;⌈(!Q && !R)⌉;⌈(!Q && R)⌉ ∧ ℓ < T;⌈(!Q && !R)⌉;⌈!Q⌉;⌈Q⌉;true',
          'True;⌈(P & (! Q))⌉;⌈(! Q)⌉;⌈((! Q) & (! R))⌉;⌈((! Q) & R)⌉ ∧ ℓ < T;⌈((! Q) & (! R))⌉;⌈(! Q)⌉;⌈Q⌉;True'),
 
     'duration_bound_l_after_until':
-        ({'P': Symbol('P'), 'Q': Symbol('Q'), 'R': Symbol('R'), 'T': Symbol('T', INT)},
+        ({'P': Symbol('P'), 'Q': Symbol('Q'), 'R': Symbol('R'), 'T': Symbol('T', REAL)},
          'true;⌈P⌉;⌈!Q⌉;⌈(!Q && !R)⌉;⌈(!Q && R)⌉ ∧ ℓ < T;⌈(!Q && !R)⌉;true',
          'True;⌈P⌉;⌈(! Q)⌉;⌈((! Q) & (! R))⌉;⌈((! Q) & R)⌉ ∧ ℓ < T;⌈((! Q) & (! R))⌉;True'),
 
     'response_delay_globally':
-        ({'R': Symbol('R'), 'S': Symbol('S'), 'T': Symbol('T', INT)},
+        ({'R': Symbol('R'), 'S': Symbol('S'), 'T': Symbol('T', REAL)},
          'true;⌈(R && !S)⌉;⌈!S⌉ ∧ ℓ > T;true',
          'True;⌈(R & (! S))⌉;⌈(! S)⌉ ∧ ℓ > T;True'),
 
     'response_delay_before':
-        ({'P': Symbol('P'), 'R': Symbol('R'), 'S': Symbol('S'), 'T': Symbol('T', INT)},
+        ({'P': Symbol('P'), 'R': Symbol('R'), 'S': Symbol('S'), 'T': Symbol('T', REAL)},
          '⌈!P⌉;⌈(!P && (R && !S))⌉;⌈(!P && !S)⌉ ∧ ℓ > T;true',
          '⌈(! P)⌉;⌈((! P) & (R & (! S)))⌉;⌈((! P) & (! S))⌉ ∧ ℓ > T;True'),
 
     'response_delay_after':
-        ({'P': Symbol('P'), 'R': Symbol('R'), 'S': Symbol('S'), 'T': Symbol('T', INT)},
+        ({'P': Symbol('P'), 'R': Symbol('R'), 'S': Symbol('S'), 'T': Symbol('T', REAL)},
          'true;⌈P⌉;true;⌈(R && !S)⌉;⌈!S⌉ ∧ ℓ > T;true',
          'True;⌈P⌉;True;⌈(R & (! S))⌉;⌈(! S)⌉ ∧ ℓ > T;True'),
 
     'response_delay_between':
-        ({'P': Symbol('P'), 'Q': Symbol('Q'), 'R': Symbol('R'), 'S': Symbol('S'), 'T': Symbol('T', INT)},
+        ({'P': Symbol('P'), 'Q': Symbol('Q'), 'R': Symbol('R'), 'S': Symbol('S'), 'T': Symbol('T', REAL)},
          'true;⌈(P && !Q)⌉;⌈!Q⌉;⌈(!Q && (R && !S))⌉;⌈(!Q && !S)⌉ ∧ ℓ > T;⌈!Q⌉;⌈Q⌉;true',
          'True;⌈(P & (! Q))⌉;⌈(! Q)⌉;⌈((! Q) & (R & (! S)))⌉;⌈((! Q) & (! S))⌉ ∧ ℓ > T;⌈(! Q)⌉;⌈Q⌉;True'),
 
     'response_delay_after_until':
-        ({'P': Symbol('P'), 'Q': Symbol('Q'), 'R': Symbol('R'), 'S': Symbol('S'), 'T': Symbol('T', INT)},
+        ({'P': Symbol('P'), 'Q': Symbol('Q'), 'R': Symbol('R'), 'S': Symbol('S'), 'T': Symbol('T', REAL)},
          'true;⌈P⌉;⌈!Q⌉;⌈(!Q && (R && !S))⌉;⌈(!Q && !S)⌉ ∧ ℓ > T;true',
          'True;⌈P⌉;⌈(! Q)⌉;⌈((! Q) & (R & (! S)))⌉;⌈((! Q) & (! S))⌉ ∧ ℓ > T;True')
 }
@@ -113,23 +113,23 @@ class TestCounterTrace(TestCase):
     '''
     @parameterized.expand([
         ('BndResponsePatternUT', 'GLOBALLY',
-         {'R': Symbol('R'), 'S': Symbol('S'), 'T': Symbol('T', INT)},
+         {'R': Symbol('R'), 'S': Symbol('S'), 'T': Symbol('T', REAL)},
          'True;⌈(R & (! S))⌉;⌈(! S)⌉ ∧ ℓ > T;True'),
 
         ('BndResponsePatternUT', 'BEFORE',
-         {'P': Symbol('P'), 'R': Symbol('R'), 'S': Symbol('S'), 'T': Symbol('T', INT)},
+         {'P': Symbol('P'), 'R': Symbol('R'), 'S': Symbol('S'), 'T': Symbol('T', REAL)},
          '⌈(! P)⌉;⌈(((! P) & R) & (! S))⌉;⌈((! P) & (! S))⌉ ∧ ℓ > T;True'),
 
         ('BndResponsePatternUT', 'AFTER',
-         {'P': Symbol('P'), 'R': Symbol('R'), 'S': Symbol('S'), 'T': Symbol('T', INT)},
+         {'P': Symbol('P'), 'R': Symbol('R'), 'S': Symbol('S'), 'T': Symbol('T', REAL)},
          'True;⌈P⌉;True;⌈(R & (! S))⌉;⌈(! S)⌉ ∧ ℓ > T;True'),
 
         ('BndResponsePatternUT', 'BETWEEN',
-         {'P': Symbol('P'), 'Q': Symbol('Q'), 'R': Symbol('R'), 'S': Symbol('S'), 'T': Symbol('T', INT)},
+         {'P': Symbol('P'), 'Q': Symbol('Q'), 'R': Symbol('R'), 'S': Symbol('S'), 'T': Symbol('T', REAL)},
          'True;⌈(P & (! Q))⌉;⌈(! Q)⌉;⌈(((! Q) & R) & (! S))⌉;⌈((! Q) & (! S))⌉ ∧ ℓ > T;⌈(! Q)⌉;⌈Q⌉;True'),
 
         ('BndResponsePatternUT', 'AFTER_UNTIL',
-         {'P': Symbol('P'), 'Q': Symbol('Q'), 'R': Symbol('R'), 'S': Symbol('S'), 'T': Symbol('T', INT)},
+         {'P': Symbol('P'), 'Q': Symbol('Q'), 'R': Symbol('R'), 'S': Symbol('S'), 'T': Symbol('T', REAL)},
          'True;⌈P⌉;⌈(! Q)⌉;⌈(((! Q) & R) & (! S))⌉;⌈((! Q) & (! S))⌉ ∧ ℓ > T;True'),
     ])
     def test_bnd_response_pattern_ut(self, pattern: str, scope: str, expressions: dict[str, FNode], expected: str):
